Merged PR 63017: Display version on dashboard for Checkout Sms
## What's being changed

We have added the Hyva module version to the Dotdigital dashboard.

## Why it's being changed

We have had many queries where we needed to ask if the module is installed or if they have the correct versions installed via support. This change will allow support to check the version themselves without having to ask the merchant's developer to run composer commands.

## How to review / test this change

- Clear cache
- Navigate to the Dotdigital dashboard.
- Observe the Hyva module and version to confirm they are correct.

Related work items: #284772